+++ models.py
@@ -1,0 +1,4 @@
+from pydantic import BaseModel
+
+class AuditResponses(BaseModel):
+    responses: dict[str, str]
