+++ requirements.txt
@@ -1,5 +1,7 @@
-Flask==2.2.3
-Werkzeug==2.2.3
-requests==2.28.2
-beautifulsoup4==4.12.2
-gunicorn==20.1.0
+fastapi==0.111.0
+uvicorn==0.30.1
+requests==2.32.3
+pydantic==2.8.2
+pydantic-settings==2.3.4
+python-dotenv==1.0.1
+openai
